pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-hospital-scheduler"
version = "0.1.0"
description = "Hospital Bed & Staff Scheduler environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # OpenAI client used by inference.py for LLM-based agent
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
server = "hospital_scheduler.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["hospital_scheduler", "hospital_scheduler.server"]
package-dir = { "hospital_scheduler" = ".", "hospital_scheduler.server" = "server" }